fix: change /apps-check-permission featureCode to long type

+++ src/main/java/org/ukdw/dto/request/auth/AppsCheckPermissionRequest.java
@@ -18,5 +18,5 @@ import lombok.Setter;
 @Getter
 public class AppsCheckPermissionRequest {
     @NotNull(message = "feature code is required")
-    private int featureCode;
+    private long featureCode;
 }
